Refactor: SideNavigationLayout 컴포넌트 -> 페이지 변경#51
